requirements.txt
pillow==10.4.0
python-docx==1.1.2